+++ flowless-sample-single-root/src/main/java/com/zhuinden/flowless_dispatcher_sample/MainActivity.java
@@ -1,9 +1,6 @@
 package com.zhuinden.flowless_dispatcher_sample;
 
 import android.content.Context;
-import android.content.Intent;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.ViewGroup;
@@ -23,7 +20,7 @@ public class MainActivity
 
     @Override
     protected void attachBaseContext(Context newBase) {
-        flowDispatcher = new SingleRootDispatcher(newBase, this);
+        flowDispatcher = new SingleRootDispatcher(this);
         newBase = Flow.configure(newBase, this) //
                 .defaultKey(FirstKey.create()) //
                 .dispatcher(flowDispatcher) //
@@ -41,30 +38,11 @@ public class MainActivity
     }
 
     @Override
-    protected void onPostCreate(@Nullable Bundle savedInstanceState) {
-        super.onPostCreate(savedInstanceState);
-        flowDispatcher.onActivityCreated(this, savedInstanceState); // dispatcher root cannot know about this event unless explicitly called
-        //please note that this method is called after `onStart()`. It's most likely not needed anyways.
-    }
-
-    @Override
     public void onBackPressed() {
         boolean didGoBack = flowDispatcher.onBackPressed();
         if(didGoBack) {
             return;
         }
         super.onBackPressed();
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-        flowDispatcher.onActivityResult(requestCode, resultCode, data);
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        flowDispatcher.onRequestPermissionsResult(requestCode, permissions, grantResults);
     }
 }