+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Change Log
 ==========
 
+Flowless Version 1.0-alpha2 *(2016-07-04)*
+--------------------------------
+* `onActivityResult` and `onPermissionResult` callbacks added to `InternalLifecycleIntegration` (Activity no longer has to do this manually)
+* `onCreate` callback now relies on `InternalLifecycleIntegration`'s `onActivityCreated()` callback (Activity no longer has to do this manually)
+* Dispatchers rely on the InternalLifecycleIntegration's callback events instead of being an `Application.ActivityCallbacks` (simplifies delegation in `SinglePaneContainer`)
+* Removed `FlowActivityCallbacks`, it was the same as `FlowLifecycles.BackPressListener` and the other two listeners.
+
 Flowless Version 1.0-alpha *(2016-07-01)*
 --------------------------------
 * Updated `SingleRootDispatcher` sample